LayoutTests/platform/ios/imported/w3c/web-platform-tests/editing/other/delete-in-child-of-head.tentative_designMode=off&method=forwarddelete-expected.txt



FAIL forwarddelete in <div> elements in <head> should join them assert_in_array: The <div> elements should be merged value "<head><title>iframe</title><div>abc</div><div>def</div></head><body><br></body>" not in array ["<head><title>iframe</title><div>abcdef</div></head><body><br></body>", "<head><title>iframe</title><div>abcdef<br></div></head><body><br></body>"]
FAIL forwarddelete around invisible <meta> should not delete it at joining paragraphs assert_in_array: The <div> elements should be merged without deleting <meta> value "<head><title>iframe</title><div>abc</div><meta><div>def</div></head><body><br></body>" not in array ["<head><title>iframe</title><div>abcdef</div><meta></head><body><br></body>", "<head><title>iframe</title><div>abcdef<br></div><meta></head><body><br></body>", "<head><title>iframe</title><meta><div>abcdef</div></head><body><br></body>", "<head><title>iframe</title><meta><div>abcdef<br></div></head><body><br></body>"]
FAIL forwarddelete around invisible <title> should not delete it at joining paragraphs assert_in_array: The <div> elements should be merged without deleting <title> value "<head><title>iframe</title><div>abc</div><title></title><div>def</div></head><body><br></body>" not in array ["<head><title>iframe</title><div>abcdef</div><title></title></head><body><br></body>", "<head><title>iframe</title><div>abcdef<br></div><title></title></head><body><br></body>", "<head><title>iframe</title><title></title><div>abcdef</div></head><body><br></body>", "<head><title>iframe</title><title></title><div>abcdef<br></div></head><body><br></body>"]
FAIL forwarddelete around invisible <style> should not delete it at joining paragraphs assert_in_array: The <div> elements should be merged without deleting <style> value "<head><title>iframe</title><div>abc</div><style></style><div>def</div></head><body><br></body>" not in array ["<head><title>iframe</title><div>abcdef</div><style></style></head><body><br></body>", "<head><title>iframe</title><div>abcdef<br></div><style></style></head><body><br></body>", "<head><title>iframe</title><style></style><div>abcdef</div></head><body><br></body>", "<head><title>iframe</title><style></style><div>abcdef<br></div></head><body><br></body>"]
FAIL forwarddelete around invisible <script> should not delete it at joining paragraphs assert_in_array: The <div> elements should be merged without deleting <script> value "<head><title>iframe</title><div>abc</div><script></script><div>def</div></head><body><br></body>" not in array ["<head><title>iframe</title><div>abcdef</div><script></script></head><body><br></body>", "<head><title>iframe</title><div>abcdef<br></div><script></script></head><body><br></body>", "<head><title>iframe</title><script></script><div>abcdef</div></head><body><br></body>", "<head><title>iframe</title><script></script><div>abcdef<br></div></head><body><br></body>"]
FAIL forwarddelete around invisible <link> should not delete it at joining paragraphs assert_in_array: The <div> elements should be merged without deleting <link> value "<head><title>iframe</title><div>abc</div><link><div>def</div></head><body><br></body>" not in array ["<head><title>iframe</title><div>abcdef</div><link></head><body><br></body>", "<head><title>iframe</title><div>abcdef<br></div><link></head><body><br></body>", "<head><title>iframe</title><link><div>abcdef</div></head><body><br></body>", "<head><title>iframe</title><link><div>abcdef<br></div></head><body><br></body>"]
FAIL forwarddelete around invisible <base> should not delete it at joining paragraphs assert_in_array: The <div> elements should be merged without deleting <base> value "<head><title>iframe</title><div>abc</div><base><div>def</div></head><body><br></body>" not in array ["<head><title>iframe</title><div>abcdef</div><base></head><body><br></body>", "<head><title>iframe</title><div>abcdef<br></div><base></head><body><br></body>", "<head><title>iframe</title><base><div>abcdef</div></head><body><br></body>", "<head><title>iframe</title><base><div>abcdef<br></div></head><body><br></body>"]
FAIL forwarddelete around invisible <template> should not delete it at joining paragraphs assert_in_array: The <div> elements should be merged without deleting <template> value "<head><title>iframe</title><div>abc</div><template></template><div>def</div></head><body><br></body>" not in array ["<head><title>iframe</title><div>abcdef</div><template></template></head><body><br></body>", "<head><title>iframe</title><div>abcdef<br></div><template></template></head><body><br></body>", "<head><title>iframe</title><template></template><div>abcdef</div></head><body><br></body>", "<head><title>iframe</title><template></template><div>abcdef<br></div></head><body><br></body>"]
PASS forwarddelete in visible <script> elements in <head> should not join them
PASS forwarddelete in visible <style> elements in <head> should not join them
PASS forwarddelete in visible <title> elements in <head> should not join them
PASS forwarddelete at boundary of <script> and <div> in <head> should not join them
PASS forwarddelete at boundary of <style> and <div> in <head> should not join them
FAIL forwarddelete at boundary of <title> and <div> in <head> should not join them assert_equals: Visible <title> and <div> shouldn't be merged expected "<head><title>iframe</title><title>abc</title><div>def</div></head><body><br></body>" but got "<head><title>iframe</title><title title=\"display:block\">abc</title><div>def</div></head><body><br></body>"
PASS forwarddelete at boundary of <div> and <script> in <head> should not join them
PASS forwarddelete at boundary of <div> and <style> in <head> should not join them
PASS forwarddelete at boundary of <div> and <title> in <head> should not join them
